msr_automation_bundle/code/requirements.txt
pandas>=2.1
openpyxl>=3.1
XlsxWriter>=3.1
PyYAML>=6.0
matplotlib>=3.8
